backend/requirements.txt
flask
flask-Cors
flask-RESTful
mysql-connector-python
gunicorn
werkzeug